pyproject.toml
[build-system]
requires = ["setuptools>=66.0.0", "wheel>=0.40.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rna-predict"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

# Core dependencies
dependencies = [
    "biopython>=1.83",
    "black>=25.1.0",
    "datasets>=3.3.2",
    "einops>=0.8.1",
    "hypothesis[cli]>=6.130.4",
    "isort>=6.0.1",
    "lxml>=5.3.1",
    "mdanalysis>=2.9.0",
    "protenix>=0.4.4",
    "ruff>=0.11.2",
    "torch>=2.0.1, <2.6.0",
    "transformers>=4.49.0",
    "pyautogui>=0.9.54",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "mss>=9.0.1",
    "dearpygui>=1.10",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "cosmic-ray>=8.4.1,<9.0.0",
]

# If you have optional (dev) dependencies:
[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "coverage>=7.6.12",
    "flake8>=7.1.2",
    "gitchangelog>=3.0.4",
    "isort>=6.0.1",
    "mkdocs>=1.6.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "pymdown-extensions",
]
gpu = [
    "nvidia-cublas-cu12==12.1.3.1; sys_platform=='linux'"
]

# If you want a console script:
[project.scripts]
rna_predict = "rna_predict.__main__:main"

[dependency-groups]
dev = [
    "cosmic-ray>=8.4.1",
    "mkdocs>=1.6.1",
    "mutatest>=2.0.1",
    "pydeps>=3.0.1",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-faulthandler>=2.0.1",
    "pytest-memprof>=0.2.0",
    "pytest-timeout>=2.3.1",
    "snoop>=0.6.0",
    "types-requests>=2.32.0.20250328",
]

[tool.setuptools.packages.find]
include = ["rna_predict*"]
exclude = ["site"]

[tool.uv]
override-dependencies = [
    "nvidia-cublas-cu12==12.1.3.1; sys_platform=='linux'"
]

# Add pytest settings to fix warnings and set fixture loop scope
[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"


# Mypy configuration
[tool.mypy]
# Ignore missing stubs for specific libraries
[[tool.mypy.overrides]]
module = "cv2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dearpygui.*"
ignore_missing_imports = true

